Status Led refactor
