+++ .gitignore
@@ -5,4 +5,4 @@ logs/
 CMakeListsPrivate.txt
 cmake-build-*/ 
 secrets.h
-src/www/*.json
+www/*.json
