+++ .idea/encodings.xml
@@ -2,6 +2,6 @@
 <project version="4">
   <component name="Encoding">
     <file url="file://$PROJECT_DIR$/include/pixel_js.h" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/src/www/pixel.js" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/www/pixel.js" charset="UTF-8" />
   </component>
 </project>
